+++ src/utils/binary.cpp
@@ -5,25 +5,35 @@
  * \author xythobuz
  */
 
+#include "global.h"
 #include "utils/binary.h"
 
-BinaryFile::BinaryFile(const char *f) {
-    file.open(f, std::ios_base::in | std::ios_base::binary);
+BinaryFile::~BinaryFile() {
+    if (file.is_open())
+        file.close();
 }
 
-BinaryFile::~BinaryFile() {
-    file.close();
+int BinaryFile::open(const char *f) {
+    if (file.is_open()) {
+        return 1;
+    } else {
+        file.open(f, std::ios_base::in | std::ios_base::binary);
+        return (file ? 0 : 1);
+    }
 }
 
 long long BinaryFile::tell() {
+    assert(file.is_open());
     return file.tellg();
 }
 
 void BinaryFile::seek(long long pos) {
+    assert(file.is_open());
     file.seekg(pos);
 }
 
 uint8_t BinaryFile::readU8() {
+    assert(file.is_open());
     uint8_t ret;
     char *c = reinterpret_cast<char *>(&ret);
     file.read(c, 1);
@@ -31,24 +41,28 @@ uint8_t BinaryFile::readU8() {
 }
 
 uint16_t BinaryFile::readU16() {
+    assert(file.is_open());
     uint8_t a = readU8();
     uint8_t b = readU8();
     return ((uint16_t)a | (uint16_t)(b << 8));
 }
 
 uint32_t BinaryFile::readU32() {
+    assert(file.is_open());
     uint16_t a = readU16();
     uint16_t b = readU16();
     return ((uint32_t)a | (uint32_t)(b << 16));
 }
 
 uint64_t BinaryFile::readU64() {
+    assert(file.is_open());
     uint32_t a = readU32();
     uint32_t b = readU32();
     return ((uint64_t)a | ((uint64_t)b << 32));
 }
 
 float BinaryFile::readFloat() {
+    assert(file.is_open());
     uint32_t val = readU32();
     char *a = reinterpret_cast<char *>(&val);
 
@@ -82,6 +96,7 @@ namespace {
 }
 
 int8_t BinaryFile::read8() {
+    assert(file.is_open());
     int8_t ret;
     char *p = reinterpret_cast<char *>(&ret);
     file.read(p, sizeof(ret));
@@ -89,6 +104,7 @@ int8_t BinaryFile::read8() {
 }
 
 int16_t BinaryFile::read16() {
+    assert(file.is_open());
     int16_t ret;
     char *p = reinterpret_cast<char *>(&ret);
     file.read(p, sizeof(ret));
@@ -97,6 +113,7 @@ int16_t BinaryFile::read16() {
 }
 
 int32_t BinaryFile::read32() {
+    assert(file.is_open());
     int32_t ret;
     char *p = reinterpret_cast<char *>(&ret);
     file.read(p, sizeof(ret));
@@ -105,6 +122,7 @@ int32_t BinaryFile::read32() {
 }
 
 int64_t BinaryFile::read64() {
+    assert(file.is_open());
     int64_t ret;
     char *p = reinterpret_cast<char *>(&ret);
     file.read(p, sizeof(ret));